clean up the state
