+++ beacon-chain/state/v2/BUILD.bazel
@@ -11,8 +11,24 @@ go_library(
         "field_root_validator.go",
         "field_root_vector.go",
         "field_roots.go",
-        "getters.go",
-        "setters.go",
+        "getters_block.go",
+        "getters_checkpoint.go",
+        "getters_eth1.go",
+        "getters_misc.go",
+        "getters_participation.go",
+        "getters_randao.go",
+        "getters_state.go",
+        "getters_sync_committee.go",
+        "getters_validator.go",
+        "setters_block.go",
+        "setters_checkpoint.go",
+        "setters_eth1.go",
+        "setters_misc.go",
+        "setters_participation.go",
+        "setters_randao.go",
+        "setters_state.go",
+        "setters_sync_committee.go",
+        "setters_validator.go",
         "state_trie.go",
         "types.go",
     ],
@@ -40,6 +56,8 @@ go_library(
         "//shared/bytesutil:go_default_library",
         "@com_github_dgraph_io_ristretto//:go_default_library",
         "@com_github_pkg_errors//:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus:go_default_library",
+        "@com_github_prometheus_client_golang//prometheus/promauto:go_default_library",
         "@com_github_prysmaticlabs_eth2_types//:go_default_library",
         "@com_github_prysmaticlabs_go_bitfield//:go_default_library",
         "@io_opencensus_go//trace:go_default_library",
@@ -52,12 +70,17 @@ go_test(
     srcs = [
         "deprecated_getters_test.go",
         "deprecated_setters_test.go",
+        "getters_block_test.go",
         "getters_test.go",
+        "getters_validator_test.go",
+        "setters_test.go",
         "state_trie_test.go",
     ],
     embed = [":go_default_library"],
     deps = [
         "//beacon-chain/state/stateutil:go_default_library",
+        "//beacon-chain/state/types:go_default_library",
+        "//beacon-chain/state/v1:go_default_library",
         "//config/params:go_default_library",
         "//proto/prysm/v1alpha1:go_default_library",
         "//shared/bytesutil:go_default_library",
