+++ beacon-chain/state/v2/getters_validator_test.go
@@ -1,0 +1,20 @@
+package v2_test
+
+import (
+	"testing"
+
+	v1 "github.com/prysmaticlabs/prysm/beacon-chain/state/v1"
+	ethpb "github.com/prysmaticlabs/prysm/proto/prysm/v1alpha1"
+	"github.com/prysmaticlabs/prysm/shared/testutil/assert"
+	"github.com/prysmaticlabs/prysm/shared/testutil/require"
+)
+
+func TestBeaconState_ValidatorAtIndexReadOnly_HandlesNilSlice(t *testing.T) {
+	st, err := v1.InitializeFromProtoUnsafe(&ethpb.BeaconState{
+		Validators: nil,
+	})
+	require.NoError(t, err)
+
+	_, err = st.ValidatorAtIndexReadOnly(0)
+	assert.Equal(t, v1.ErrNilValidatorsInState, err)
+}
