+++ beacon-chain/state/v2/setters_checkpoint.go
@@ -1,0 +1,58 @@
+package v2
+
+import (
+	"github.com/prysmaticlabs/go-bitfield"
+	ethpb "github.com/prysmaticlabs/prysm/proto/prysm/v1alpha1"
+)
+
+// SetJustificationBits for the beacon state.
+func (b *BeaconState) SetJustificationBits(val bitfield.Bitvector4) error {
+	if !b.hasInnerState() {
+		return ErrNilInnerState
+	}
+	b.lock.Lock()
+	defer b.lock.Unlock()
+
+	b.state.JustificationBits = val
+	b.markFieldAsDirty(justificationBits)
+	return nil
+}
+
+// SetPreviousJustifiedCheckpoint for the beacon state.
+func (b *BeaconState) SetPreviousJustifiedCheckpoint(val *ethpb.Checkpoint) error {
+	if !b.hasInnerState() {
+		return ErrNilInnerState
+	}
+	b.lock.Lock()
+	defer b.lock.Unlock()
+
+	b.state.PreviousJustifiedCheckpoint = val
+	b.markFieldAsDirty(previousJustifiedCheckpoint)
+	return nil
+}
+
+// SetCurrentJustifiedCheckpoint for the beacon state.
+func (b *BeaconState) SetCurrentJustifiedCheckpoint(val *ethpb.Checkpoint) error {
+	if !b.hasInnerState() {
+		return ErrNilInnerState
+	}
+	b.lock.Lock()
+	defer b.lock.Unlock()
+
+	b.state.CurrentJustifiedCheckpoint = val
+	b.markFieldAsDirty(currentJustifiedCheckpoint)
+	return nil
+}
+
+// SetFinalizedCheckpoint for the beacon state.
+func (b *BeaconState) SetFinalizedCheckpoint(val *ethpb.Checkpoint) error {
+	if !b.hasInnerState() {
+		return ErrNilInnerState
+	}
+	b.lock.Lock()
+	defer b.lock.Unlock()
+
+	b.state.FinalizedCheckpoint = val
+	b.markFieldAsDirty(finalizedCheckpoint)
+	return nil
+}
