+++ beacon-chain/state/v2/state_trie.go
@@ -6,6 +6,8 @@ import (
 	"sort"
 
 	"github.com/pkg/errors"
+	"github.com/prometheus/client_golang/prometheus"
+	"github.com/prometheus/client_golang/prometheus/promauto"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state/fieldtrie"
 	"github.com/prysmaticlabs/prysm/beacon-chain/state/stateutil"
@@ -18,6 +20,13 @@ import (
 	"github.com/prysmaticlabs/prysm/shared/bytesutil"
 	"go.opencensus.io/trace"
 	"google.golang.org/protobuf/proto"
+)
+
+var (
+	stateCount = promauto.NewGauge(prometheus.GaugeOpts{
+		Name: "beacon_state_altair_count",
+		Help: "Count the number of active beacon state objects.",
+	})
 )
 
 // InitializeFromProto the beacon state from a protobuf representation.
@@ -35,7 +44,7 @@ func InitializeFromProtoUnsafe(st *ethpb.BeaconStateAltair) (*BeaconState, error
 	fieldCount := params.BeaconConfig().BeaconStateAltairFieldCount
 	b := &BeaconState{
 		state:                 st,
-		dirtyFields:           make(map[types.FieldIndex]interface{}, fieldCount),
+		dirtyFields:           make(map[types.FieldIndex]bool, fieldCount),
 		dirtyIndices:          make(map[types.FieldIndex][]uint64, fieldCount),
 		stateFieldLeaves:      make(map[types.FieldIndex]*fieldtrie.FieldTrie, fieldCount),
 		sharedFieldReferences: make(map[types.FieldIndex]*stateutil.Reference, 11),
@@ -67,6 +76,7 @@ func InitializeFromProtoUnsafe(st *ethpb.BeaconStateAltair) (*BeaconState, error
 	b.sharedFieldReferences[inactivityScores] = stateutil.NewRef(1) // New in Altair.
 	b.sharedFieldReferences[historicalRoots] = stateutil.NewRef(1)
 
+	stateCount.Inc()
 	return b, nil
 }
 
@@ -114,13 +124,13 @@ func (b *BeaconState) Copy() state.BeaconState {
 			CurrentSyncCommittee:        b.currentSyncCommittee(),
 			NextSyncCommittee:           b.nextSyncCommittee(),
 		},
-		dirtyFields:           make(map[types.FieldIndex]interface{}, fieldCount),
+		dirtyFields:           make(map[types.FieldIndex]bool, fieldCount),
 		dirtyIndices:          make(map[types.FieldIndex][]uint64, fieldCount),
 		rebuildTrie:           make(map[types.FieldIndex]bool, fieldCount),
 		sharedFieldReferences: make(map[types.FieldIndex]*stateutil.Reference, 11),
 		stateFieldLeaves:      make(map[types.FieldIndex]*fieldtrie.FieldTrie, fieldCount),
 
-		// Copy on write validator index map.
+		// Share the reference to validator index map.
 		valMapHandler: b.valMapHandler,
 	}
 
@@ -166,6 +176,7 @@ func (b *BeaconState) Copy() state.BeaconState {
 		}
 	}
 
+	stateCount.Inc()
 	// Finalizer runs when dst is being destroyed in garbage collection.
 	runtime.SetFinalizer(dst, func(b *BeaconState) {
 		for field, v := range b.sharedFieldReferences {
@@ -174,6 +185,15 @@ func (b *BeaconState) Copy() state.BeaconState {
 				b.stateFieldLeaves[field].FieldReference().MinusRef()
 			}
 		}
+		for i := 0; i < fieldCount; i++ {
+			field := types.FieldIndex(i)
+			delete(b.stateFieldLeaves, field)
+			delete(b.dirtyIndices, field)
+			delete(b.dirtyFields, field)
+			delete(b.sharedFieldReferences, field)
+			delete(b.stateFieldLeaves, field)
+		}
+		stateCount.Sub(1)
 	})
 
 	return dst
@@ -189,17 +209,17 @@ func (b *BeaconState) HashTreeRoot(ctx context.Context) ([32]byte, error) {
 	defer b.lock.Unlock()
 
 	if b.merkleLayers == nil || len(b.merkleLayers) == 0 {
-		fieldRoots, err := computeFieldRoots(b.state)
+		fieldRoots, err := computeFieldRoots(ctx, b.state)
 		if err != nil {
 			return [32]byte{}, err
 		}
 		layers := stateutil.Merkleize(fieldRoots)
 		b.merkleLayers = layers
-		b.dirtyFields = make(map[types.FieldIndex]interface{}, params.BeaconConfig().BeaconStateAltairFieldCount)
+		b.dirtyFields = make(map[types.FieldIndex]bool, params.BeaconConfig().BeaconStateAltairFieldCount)
 	}
 
 	for field := range b.dirtyFields {
-		root, err := b.rootSelector(field)
+		root, err := b.rootSelector(ctx, field)
 		if err != nil {
 			return [32]byte{}, err
 		}
@@ -236,7 +256,11 @@ func (b *BeaconState) IsNil() bool {
 	return b == nil || b.state == nil
 }
 
-func (b *BeaconState) rootSelector(field types.FieldIndex) ([32]byte, error) {
+func (b *BeaconState) rootSelector(ctx context.Context, field types.FieldIndex) ([32]byte, error) {
+	ctx, span := trace.StartSpan(ctx, "beaconState.rootSelector")
+	defer span.End()
+	span.AddAttributes(trace.StringAttribute("field", field.String(b.Version())))
+
 	hasher := hash.CustomSHA256Hasher()
 	switch field {
 	case genesisTime:
@@ -257,7 +281,6 @@ func (b *BeaconState) rootSelector(field types.FieldIndex) ([32]byte, error) {
 			if err != nil {
 				return [32]byte{}, err
 			}
-			b.dirtyIndices[field] = []uint64{}
 			delete(b.rebuildTrie, field)
 			return b.stateFieldLeaves[field].TrieRoot()
 		}
@@ -268,7 +291,6 @@ func (b *BeaconState) rootSelector(field types.FieldIndex) ([32]byte, error) {
 			if err != nil {
 				return [32]byte{}, err
 			}
-			b.dirtyIndices[field] = []uint64{}
 			delete(b.rebuildTrie, field)
 			return b.stateFieldLeaves[field].TrieRoot()
 		}
@@ -279,11 +301,14 @@ func (b *BeaconState) rootSelector(field types.FieldIndex) ([32]byte, error) {
 		return eth1Root(hasher, b.state.Eth1Data)
 	case eth1DataVotes:
 		if b.rebuildTrie[field] {
-			err := b.resetFieldTrie(field, b.state.Eth1DataVotes, uint64(params.BeaconConfig().SlotsPerEpoch.Mul(uint64(params.BeaconConfig().EpochsPerEth1VotingPeriod))))
+			err := b.resetFieldTrie(
+				field,
+				b.state.Eth1DataVotes,
+				uint64(params.BeaconConfig().SlotsPerEpoch.Mul(uint64(params.BeaconConfig().EpochsPerEth1VotingPeriod))),
+			)
 			if err != nil {
 				return [32]byte{}, err
 			}
-			b.dirtyIndices[field] = []uint64{}
 			delete(b.rebuildTrie, field)
 			return b.stateFieldLeaves[field].TrieRoot()
 		}
@@ -294,7 +319,6 @@ func (b *BeaconState) rootSelector(field types.FieldIndex) ([32]byte, error) {
 			if err != nil {
 				return [32]byte{}, err
 			}
-			b.dirtyIndices[validators] = []uint64{}
 			delete(b.rebuildTrie, validators)
 			return b.stateFieldLeaves[field].TrieRoot()
 		}
@@ -307,7 +331,6 @@ func (b *BeaconState) rootSelector(field types.FieldIndex) ([32]byte, error) {
 			if err != nil {
 				return [32]byte{}, err
 			}
-			b.dirtyIndices[field] = []uint64{}
 			delete(b.rebuildTrie, field)
 			return b.stateFieldLeaves[field].TrieRoot()
 		}
@@ -338,13 +361,16 @@ func (b *BeaconState) rootSelector(field types.FieldIndex) ([32]byte, error) {
 
 func (b *BeaconState) recomputeFieldTrie(index types.FieldIndex, elements interface{}) ([32]byte, error) {
 	fTrie := b.stateFieldLeaves[index]
+	// We can't lock the trie directly because the trie's variable gets reassigned,
+	// and therefore we would call Unlock() on a different object.
+	fTrieMutex := fTrie.RWMutex
 	if fTrie.FieldReference().Refs() > 1 {
-		fTrie.Lock()
-		defer fTrie.Unlock()
+		fTrieMutex.Lock()
 		fTrie.FieldReference().MinusRef()
 		newTrie := fTrie.CopyTrie()
 		b.stateFieldLeaves[index] = newTrie
 		fTrie = newTrie
+		fTrieMutex.Unlock()
 	}
 	// remove duplicate indexes
 	b.dirtyIndices[index] = slice.SetUint64(b.dirtyIndices[index])
