+++ beacon-chain/state/v2/types.go
@@ -24,6 +24,28 @@ func init() {
 	fieldMap[types.Validators] = types.CompositeArray
 }
 
+// fieldMap keeps track of each field
+// to its corresponding data type.
+var fieldMap map[types.FieldIndex]types.DataType
+
+// ErrNilInnerState returns when the inner state is nil and no copy set or get
+// operations can be performed on state.
+var ErrNilInnerState = errors.New("nil inner state")
+
+// BeaconState defines a struct containing utilities for the eth2 chain state, defining
+// getters and setters for its respective values and helpful functions such as HashTreeRoot().
+type BeaconState struct {
+	state                 *ethpb.BeaconStateAltair
+	lock                  sync.RWMutex
+	dirtyFields           map[types.FieldIndex]bool
+	dirtyIndices          map[types.FieldIndex][]uint64
+	stateFieldLeaves      map[types.FieldIndex]*fieldtrie.FieldTrie
+	rebuildTrie           map[types.FieldIndex]bool
+	valMapHandler         *stateutil.ValidatorMapHandler
+	merkleLayers          [][][]byte
+	sharedFieldReferences map[types.FieldIndex]*stateutil.Reference
+}
+
 // Field Aliases for values from the types package.
 const (
 	genesisTime                    = types.GenesisTime
@@ -51,25 +73,3 @@ const (
 	currentSyncCommittee           = types.CurrentSyncCommittee
 	nextSyncCommittee              = types.NextSyncCommittee
 )
-
-// fieldMap keeps track of each field
-// to its corresponding data type.
-var fieldMap map[types.FieldIndex]types.DataType
-
-// ErrNilInnerState returns when the inner state is nil and no copy set or get
-// operations can be performed on state.
-var ErrNilInnerState = errors.New("nil inner state")
-
-// BeaconState defines a struct containing utilities for the eth2 chain state, defining
-// getters and setters for its respective values and helpful functions such as HashTreeRoot().
-type BeaconState struct {
-	state                 *ethpb.BeaconStateAltair
-	lock                  sync.RWMutex
-	dirtyFields           map[types.FieldIndex]interface{}
-	dirtyIndices          map[types.FieldIndex][]uint64
-	stateFieldLeaves      map[types.FieldIndex]*fieldtrie.FieldTrie
-	rebuildTrie           map[types.FieldIndex]bool
-	valMapHandler         *stateutil.ValidatorMapHandler
-	merkleLayers          [][][]byte
-	sharedFieldReferences map[types.FieldIndex]*stateutil.Reference
-}
